Moved AlertDialog inside of Trash component; js_only.
